fix int to size_t

+++ benchmark/benchmark_maximum_projections.cpp
@@ -24,9 +24,9 @@ protected:
         std::vector<float> inputData (dataWidth * dataWidth * dataWidth);
 
         // Initialise device memory and push from host
-        std::array<int, 3> dimensionsInput = {dataWidth, dataWidth, dataWidth};
+        std::array<size_t, 3> dimensionsInput = {dataWidth, dataWidth, dataWidth};
         gpuInput = cle.Push<float>(inputData, dimensionsInput);
-        std::array<int, 3> dimensionsOutput = {dataWidth, dataWidth, 1};
+        std::array<size_t, 3> dimensionsOutput = {dataWidth, dataWidth, 1};
         gpuOutput = cle.Create<float>(dimensionsOutput);
     }
 
@@ -35,7 +35,7 @@ protected:
     virtual void Teardown(){};
 
 public:
-    int dataWidth;
+    size_t dataWidth;
     MaximumProjectionBenchmarkBase() : cle(cle::Clesperanto()){}
     virtual ~MaximumProjectionBenchmarkBase(){}
 };
