+++ benchmark/benchmark_means_comparison.cpp
@@ -25,14 +25,14 @@ protected:
     {
         vector<float> inputData(dataWidth * dataWidth);
         
-        array<int,3> dim{{dataWidth, dataWidth, 1}};
+        array<size_t,3> dim{{dataWidth, dataWidth, 1}};
         gpuInput = cle.Push<float>(inputData, dim);
         gpuOutput = cle.Create<float>(dim);
     }
     virtual void Teardown() {}
 
 public:
-    int dataWidth;
+    size_t dataWidth;
     int radius = 3;
     MeanBoxBenchmark(const cle::Clesperanto& _cle) : cle(_cle){}
     MeanBoxBenchmark() : cle(cle::Clesperanto()){}
@@ -49,7 +49,7 @@ protected:
 
     virtual void Compile(cle::Clesperanto& cle)
     {
-        array<int,3> dim{{1, 1, 1}};
+        array<size_t,3> dim{{1, 1, 1}};
         cle::Buffer in = cle.Create<float>(dim);
         cle::Buffer out = cle.Create<float>(dim);
         cle.MeanSphere(in, out, 1, 1);
@@ -72,7 +72,7 @@ protected:
 
     virtual void Compile(cle::Clesperanto& cle)
     {
-        array<int,3> dim{{1, 1, 1}};
+        array<size_t,3> dim{{1, 1, 1}};
         cle::Buffer in = cle.Create<float>(dim);
         cle::Buffer out = cle.Create<float>(dim);
         cle.MeanBox(in, out, 1, 1);
